+++ core/class/AbeilleParser.class.php
@@ -4285,31 +4285,31 @@
                         }
 
                         // Ikea specific ?
-                        else if ($cmd == "07") {
-                            $manufCode = substr($payload,30, 2).substr($payload,28, 2);
-                            if ( $manufCode=='117C' ) {
+                        else if (($cmd == "07") && ($manufCode == '117C')) {
+                            // $manufCode = substr($payload,30, 2).substr($payload,28, 2);
+                            // if (  ) {
                                 // $sqn                    = substr($payload,32, 2);
                                 // $cmd                    = substr($payload,34, 2);
                                 // if ( $cmd != "07" ) {
                                 //     parserLog("debug", '  Message can t be decoded. Looks like Telecommande Ikea Ronde but not completely.');
                                 //     return;
                                 // }
-                                $remainingData = substr($payload,36, 8);
-                                $value = substr($payload,36, 2);
+                                // $remainingData = substr($payload,36, 8);
+                                $value = substr($pl, 0, 2);
 
-                                parserLog("debug", '  Telecommande Ikea Ronde'
+                                parserLog("debug", '  Ikea 5 buttons remote'
                                                .', FCF='.$frameCtrlField
                                                .', ManufCode='.$manufCode
                                                .', SQN='.$sqn
-                                               .', cmd='.$cmd
-                                               .', value='.$value
+                                               .', Cmd='.$cmd
+                                               .', Value='.$value
                                                 );
 
-                                $attrReportN = [
+                                $attrReportN[] = [
                                     array( "name" => $clustId.'-'.$srcEp.'-0000', "value" => $value ),
                                 ];
                                 // return;
-                            }
+                            // }
                         }
                     } // End $clustId == "0005"